fp_16_v1

+++ config/experiment_config.py
@@ -16,7 +16,7 @@ def read_cfg(mod="non_contrastive"):
     FLAGS.Encoder_block_channel_output = {'1':1,'2':1,'3':1,'4':1,'5':1}
     
     FLAGS.loss_type ="symmetrized"# asymmetrized (2 only options)
-    FLAGS.mixprecision='fp32' #['fp16', 'fp32'],  # fp32 is original precision
+    FLAGS.mixprecision='fp16' #['fp16', 'fp32'],  # fp32 is original precision
     FLAGS.base_lr = 0.5
 
     FLAGS.non_contrast_binary_loss = "sum_symetrize_l2_loss_object_backg_add_original"#sum_symetrize_l2_loss_object_backg_add_original
